+++ app/(tabs)/index.tsx
@@ -1,14 +1,37 @@
 import { StatusBar } from 'expo-status-bar';
-import React, { useRef } from 'react';
+import React, {useEffect, useState, useRef } from 'react';
 import { Animated, InteractionManager } from 'react-native';
-import { StyleSheet, Text, View, TouchableOpacity, SafeAreaView, TextInput, Image } from 'react-native';
+import {Button, StyleSheet, Text, View, TouchableOpacity, SafeAreaView, TextInput, Image } from 'react-native';
 import AntDesign from '@expo/vector-icons/AntDesign';
 import FontAwesome from '@expo/vector-icons/FontAwesome';
 import { Colors } from 'react-native/Libraries/NewAppScreen';
 import { interpolateColor } from 'react-native-reanimated';
 import FavoriteButton from '../../components/FavoriteButton';
+import { Audio } from 'expo-av';
 
 export default function App() {
+  // 初期化
+  const [sound, setSound] = useState<Audio.Sound | null>(null);
+  useEffect(() => {
+    // アンロード
+    return sound
+      ? () => {
+        sound.unloadAsync(); 
+      }
+    : undefined;
+  }, [sound]);
+
+  // オーディオの再生
+  async function playSound() {
+    const { sound } = await Audio.Sound.createAsync(
+    require('../../assets/music/mondo_01.mp3')
+  );
+    setSound(sound);
+
+    await sound.playAsync(); 
+  }
+
+
   return (
     <SafeAreaView style={styles.safeArea}>
       <View style={styles.nonView}>
@@ -47,15 +70,15 @@ export default function App() {
 
       <View style={styles.container}>
         <View style={styles.buttonWrapper}>
-          <TouchableOpacity style={styles.musicButton}>
+          {/* <TouchableOpacity style={styles.musicButton}>
             <AntDesign name="banckward" size={40} color="white" style={styles.forwardIcon}/>
+          </TouchableOpacity> */}
+          <TouchableOpacity style={styles.musicButton} onPress={playSound}>
+            <Image style={{width: 76, height: 76}} source={require('../../assets/images/playButton1.png')}/>
           </TouchableOpacity>
-          <TouchableOpacity style={styles.musicButton}>
-            <Image style={{width: 76, height: 76}}source={require('../../assets/images/playButton1.png')}/>
-          </TouchableOpacity>
-          <TouchableOpacity style={styles.musicButton}>
+          {/* <TouchableOpacity style={styles.musicButton}>
             <AntDesign name="forward" size={40} color="white" style={styles.forwardIcon}/>
-          </TouchableOpacity>
+          </TouchableOpacity> */}
         </View>
       </View>
     </SafeAreaView>
