+++ package.json
@@ -20,6 +20,7 @@
     "@react-navigation/native": "^6.0.2",
     "aws-amplify": "^6.6.0",
     "expo": "~51.0.28",
+    "expo-av": "^14.0.7",
     "expo-constants": "~16.0.2",
     "expo-font": "~12.0.9",
     "expo-linking": "~6.3.1",
@@ -35,7 +36,7 @@
     "react-native-reanimated": "~3.10.1",
     "react-native-safe-area-context": "4.10.5",
     "react-native-screens": "3.31.1",
-    "react-native-svg": "^15.5.0",
+    "react-native-svg": "^15.2.0",
     "react-native-vector-icons": "^10.1.0",
     "react-native-web": "~0.19.10"
   },
@@ -51,10 +52,10 @@
     "constructs": "^10.3.0",
     "esbuild": "^0.23.1",
     "jest": "^29.2.1",
-    "jest-expo": "~51.0.3",
+    "jest-expo": "~51.0.4",
     "react-test-renderer": "18.2.0",
     "tsx": "^4.19.0",
-    "typescript": "^5.6.2"
+    "typescript": "~5.3.3"
   },
   "private": true
 }
